# BUG: Fix return type of loc/iloc