--- a/pandas/core/indexing.py
+++ b/pandas/core/indexing.py
@@ -28,6 +28,7 @@ from pandas.util._decorators import doc
 
 from pandas.core.dtypes.cast import (
     can_hold_element,
+    find_common_type,
     maybe_promote,
 )
 from pandas.core.dtypes.common import (
@@ -1095,14 +1096,12 @@ class _LocationIndexer(NDFrameIndexerBase):
                     return section
                 # This is an elided recursive call to iloc/loc
                 out = getattr(section, self.name)[new_key]
-                # Re-interpret dtype of out.values for loc/iloc[int, list/slice].
+                # Re-interpret dtype of out.values for loc/iloc[int, list-like].
                 # GH60600
-                if (
-                    i == 0
-                    and isinstance(key, int)
-                    and isinstance(new_key, (list, slice))
-                ):
-                    out = out.infer_objects()
+                if i == 0 and isinstance(key, int) and is_list_like(tup[1]):
+                    dt = self.obj.dtypes.__getitem__(tup[1])
+                    if len(dt) > 0:
+                        out = out.astype(find_common_type(dt.tolist()))
                 return out
 
         raise IndexingError("not applicable")