--- a/pandas/core/indexing.py
+++ b/pandas/core/indexing.py
@@ -1068,11 +1068,7 @@ class _LocationIndexer(NDFrameIndexerBase):
 
         # Reverse tuple so that we are indexing along columns before rows
         # and avoid unintended dtype inference. # GH60600
-        if any(isinstance(ax, MultiIndex) for ax in self.obj.axes):
-            enum = enumerate(tup)
-        else:
-            enum = zip(range(len(tup) - 1, -1, -1), reversed(tup))
-        for i, key in enum:
+        for i, key in zip(range(len(tup) - 1, -1, -1), reversed(tup)):
             if is_label_like(key) or is_list_like(key):
                 # We don't need to check for tuples here because those are
                 #  caught by the _is_nested_tuple_indexer check above.