@jbrockmendel 2025-03-06
I think the way to do it would be by changing getitem_lowerdim to reverse the order in which we do the indexing (i think the loop on L1069).  Would need to try it to see if that breaks anything.  Best guess would be MultiIndex cases (possibly their perf?)

I think a bunch of getitem_lowerdim dates back to when Panel and PanelND existed, so it may be more general than it needs to be.