@jbrockmendel 2025-04-02
> Should I revise these test, or just let the method do the indexing in a forward direction as I did in the first commit?

is the user-facing behavior in the affected tests changed?