@sanggon6107 2025-04-06
Hi @jbrockmendel ,
I suppose user will see an error that points out `column` not `Index`, when both `Index` and `column` have wrong input since we'll try to get the section based on the given `column` prior to `Index`. 
But I thought this might be kinda consistent behaviour since we do the same thing for `MultiIndex` at `_LocationIndexer._getitem_nested_tuple` as below.

https://github.com/pandas-dev/pandas/blob/5736b9647068d31fdf8673d3528cb64e35060bac/pandas/core/indexing.py#L1152-L1168


Reproducible example:
```python
df2 = pd.DataFrame([{"a":1.0, "b":2, "c":3, "d":4}]).set_index(["c", "d"])
>>>
       a  b
c d        
3 4  1.0  2

r = df.loc[("X", "Y"), "Z"]
>>>
...
in get_loc
    raise KeyError(key) from err
KeyError: 'Z'

```