@jbrockmendel 2025-04-13
Are you sure that example is right?  It looks like you define `df2` but then access `df`.