@jbrockmendel 2025-04-13
Thanks I think I understand now.  If the user-facing change is just a different exception message (and the new one still looks quite useful) then this seems reasonable to me