@sanggon6107 2025-04-17
Thanks, @rhshadrach , Here's the result from my side.

<details>

```
| Change   | Before [5fef9793] <main>   | After [3596f764] <fix-loc-dtype>   |   Ratio | Benchmark (Parameter)                                                                                                                                                 |
|----------|----------------------------|------------------------------------|---------|-----------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| +        | 151±3μs                    | 502±8μs                            |    3.32 | indexing_engines.NumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.UInt64Engine'>, <class 'numpy.uint64'>), 'monotonic_decr', False, 100000) |
| +        | 190±4μs                    | 224±10μs                           |    1.18 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.MaskedInt16Engine'>, 'Int16'), 'monotonic_decr', True, 100000)                  |
| -        | 9.13±0.8ms                 | 8.28±0.06ms                        |    0.91 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.MaskedUInt8Engine'>, 'UInt8'), 'monotonic_decr', False, 2000000)    |
| -        | 1.85±0.06ms                | 1.63±0.03ms                        |    0.88 | indexing.NumericSeriesIndexing.time_loc_array(<class 'numpy.uint64'>, 'unique_monotonic_inc')                                                                         |
| -        | 7.80±0.2ms                 | 6.89±0.07ms                        |    0.88 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.UInt64Engine'>, <class 'numpy.uint64'>), 'monotonic_decr', False, 2000000)            |
| -        | 516±20μs                   | 449±20μs                           |    0.87 | indexing.NumericMaskedIndexing.time_get_indexer_dups('UInt64', True)                                                                                                  |
| -        | 627±20μs                   | 547±10μs                           |    0.87 | indexing.NumericSeriesIndexing.time_getitem_list_like(<class 'numpy.uint64'>, 'nonunique_monotonic_inc')                                                              |
| -        | 3.82±0.2ms                 | 3.30±0.2ms                         |    0.87 | indexing.NumericSeriesIndexing.time_loc_array(<class 'numpy.float64'>, 'unique_monotonic_inc')                                                                        |
| -        | 5.66±0.5ms                 | 4.75±0.1ms                         |    0.84 | indexing.NumericSeriesIndexing.time_getitem_lists(<class 'numpy.float64'>, 'unique_monotonic_inc')                                                                    |
| -        | 10.1±0.5ms                 | 8.32±0.3ms                         |    0.83 | indexing.NumericSeriesIndexing.time_getitem_scalar(<class 'numpy.uint64'>, 'nonunique_monotonic_inc')                                                                 |
| -        | 10.0±0.3ms                 | 8.16±0.2ms                         |    0.81 | indexing.NumericSeriesIndexing.time_loc_scalar(<class 'numpy.uint64'>, 'nonunique_monotonic_inc')                                                                     |
| -        | 10.3±0.5ms                 | 8.27±0.2ms                         |    0.8  | indexing.NumericSeriesIndexing.time_loc_slice(<class 'numpy.uint64'>, 'nonunique_monotonic_inc')                                                                      |
| -        | 3.69±0.7ms                 | 2.80±0.1ms                         |    0.76 | indexing.MultiIndexing.time_loc_all_null_slices(False)                                                                                                                |
| -        | 462±9μs                    | 114±1μs                            |    0.25 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Int64Engine'>, <class 'numpy.int64'>), 'monotonic_decr', False, 100000)               |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.

```

</details>