@rhshadrach 2025-04-19
Thanks @sanggon6107 - I think we'll want to see if we can avoid the perf hit in that first row (and 2nd row if possible, but that is more okay). Or at least understand why if we cannot. I can help if desired, should have some time to look into this in the next week.