@sanggon6107 2025-04-27
Thanks for the feedback @rhshadrach !
I also wonder why we see the drop in the performance in the first row whereas it appears that there's an improvement in the last row even though they have similar parameters. I'll do some further investigation from my end and get back to you. Any help would be greatly appreciated as well.