@sanggon6107 2025-05-13
Hi @rhshadrach ,
I've additionally run the asv several times since I was not sure if `get_loc` really is related to `_getitem_lowerdim()` and the results are as follows :

<DETAILS>

```
| Change   | Before [a8113887] <main>   | After [3596f764] <fix-loc-dtype>   |   Ratio | Benchmark (Parameter)                                                                                                                                                   |
|----------|----------------------------|------------------------------------|---------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| +        | 98.7±30μs                  | 321±20μs                           |    3.25 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Int32Engine'>, <class 'numpy.int32'>), 'monotonic_decr', False, 100000)                 |
| -        | 145±2μs                    | 131±2μs                            |    0.91 | indexing.NumericSeriesIndexing.time_loc_slice(<class 'numpy.float64'>, 'nonunique_monotonic_inc')                                                                       |
| -        | 384±7μs                    | 339±7μs                            |    0.88 | indexing.CategoricalIndexIndexing.time_get_indexer_list('monotonic_decr')                                                                                               |
| -        | 1.44±0.03ms                | 1.19±0.02ms                        |    0.83 | indexing.MultiIndexing.time_loc_null_slice_plus_slice(True)                                                                                                             |
| -        | 154±40μs                   | 100±1μs                            |    0.65 | indexing_engines.NumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.Int8Engine'>, <class 'numpy.int8'>), 'monotonic_decr', False, 100000)       |
| -        | 359±10μs                   | 133±7μs                            |    0.37 | indexing_engines.NumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.Float64Engine'>, <class 'numpy.float64'>), 'monotonic_decr', False, 100000) |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```

</DETAILS>

<DETAILS>

```
| Change   | Before [a8113887] <main>   | After [3596f764] <fix-loc-dtype>   |   Ratio | Benchmark (Parameter)                                                                                                                                       |
|----------|----------------------------|------------------------------------|---------|-------------------------------------------------------------------------------------------------------------------------------------------------------------|
| +        | 189±100μs                  | 331±8μs                            |    1.75 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Float32Engine'>, <class 'numpy.float32'>), 'monotonic_decr', False, 100000) |

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE DECREASED.
```

</DETAILS>

<DETAILS>

```
| Change   | Before [a8113887] <main>   | After [3596f764] <fix-loc-dtype>   |   Ratio | Benchmark (Parameter)                                                                                                                                                   |
|----------|----------------------------|------------------------------------|---------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------------|
| +        | 83.3±2μs                   | 327±7μs                            |    3.93 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Float64Engine'>, <class 'numpy.float64'>), 'monotonic_decr', False, 100000)             |
| +        | 125±5μs                    | 365±10μs                           |    2.93 | indexing_engines.NumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.Float64Engine'>, <class 'numpy.float64'>), 'monotonic_decr', False, 100000) |
| +        | 181±100μs                  | 321±5μs                            |    1.78 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.UInt32Engine'>, <class 'numpy.uint32'>), 'monotonic_decr', False, 100000)               |
| +        | 217±100μs                  | 341±8μs                            |    1.57 | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.UInt64Engine'>, <class 'numpy.uint64'>), 'monotonic_decr', False, 100000)               |
| -        | 445±9μs                    | 404±6μs                            |    0.91 | indexing.NumericMaskedIndexing.time_get_indexer('UInt64', True)                                                                                                         |
| -        | 11.8±0.4μs                 | 10.7±0.1μs                         |    0.91 | indexing.NumericSeriesIndexing.time_getitem_scalar(<class 'numpy.int64'>, 'nonunique_monotonic_inc')                                                                    |
| -        | 221±20μs                   | 199±2μs                            |    0.9  | indexing.MultiIndexing.time_loc_partial_key_scalar(True)                                                                                                                |
| -        | 771±20μs                   | 691±20μs                           |    0.9  | indexing.NumericMaskedIndexing.time_get_indexer('Float64', True)                                                                                                        |
| -        | 344±10μs                   | 311±4μs                            |    0.9  | indexing.NumericMaskedIndexing.time_get_indexer('Int64', True)                                                                                                          |
| -        | 4.72±0.3ms                 | 4.23±0.04ms                        |    0.9  | indexing_engines.NumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.Int16Engine'>, <class 'numpy.int16'>), 'monotonic_decr', True, 2000000)                 |
| -        | 1.51±0.07ms                | 1.34±0.03ms                        |    0.89 | indexing.MultiIndexing.time_loc_all_slices(True)                                                                                                                        |
| -        | 107±3μs                    | 95.0±2μs                           |    0.89 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.MaskedUInt64Engine'>, 'UInt64'), 'non_monotonic', True, 2000000)      |
| -        | 139±4μs                    | 123±1μs                            |    0.88 | indexing.MultiIndexing.time_loc_all_scalars(True)                                                                                                                       |
| -        | 5.05±0.2ms                 | 4.42±0.03ms                        |    0.88 | indexing.NumericSeriesIndexing.time_getitem_lists(<class 'numpy.float64'>, 'unique_monotonic_inc')                                                                      |
| -        | 1.19±0.03ms                | 1.05±0.01ms                        |    0.88 | indexing.NumericSeriesIndexing.time_loc_list_like(<class 'numpy.float64'>, 'unique_monotonic_inc')                                                                      |
| -        | 62.0±3μs                   | 54.0±0.8μs                         |    0.87 | indexing.MultiIndexing.time_xs_full_key(False)                                                                                                                          |
| -        | 20.4±1μs                   | 17.5±0.3μs                         |    0.86 | indexing.NumericSeriesIndexing.time_getitem_scalar(<class 'numpy.float64'>, 'nonunique_monotonic_inc')                                                                  |
| -        | 1.78±0.04ms                | 1.51±0.01ms                        |    0.85 | indexing.NonNumericSeriesIndexing.time_getitem_list_like('datetime', 'unique_monotonic_inc')                                                                            |
| -        | 499±70μs                   | 412±10μs                           |    0.83 | indexing.NonNumericSeriesIndexing.time_getitem_list_like('string', 'non_monotonic')                                                                                     |
| -        | 13.1±2ms                   | 10.5±0.1ms                         |    0.8  | indexing_engines.MaskedNumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.MaskedFloat64Engine'>, 'Float64'), 'monotonic_decr', False, 2000000)  |
| -        | 11.7±0.3ms                 | 8.95±0.2ms                         |    0.77 | indexing.MultiIndexing.time_loc_multiindex(False)                                                                                                                       |
| -        | 16.7±2ms                   | 11.7±0.1ms                         |    0.7  | indexing.MultiIndexing.time_loc_slice_plus_null_slice(False)                                                                                                            |
| -        | 145±50μs                   | 94.5±1μs                           |    0.65 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.MaskedInt64Engine'>, 'Int64'), 'non_monotonic', True, 2000000)        |
| -        | 30.8±3μs                   | 18.8±0.6μs                         |    0.61 | indexing.NonNumericSeriesIndexing.time_getitem_pos_slice('string', 'non_monotonic')                                                                                     |
| -        | 31.9±1μs                   | 19.4±0.3μs                         |    0.61 | indexing.NonNumericSeriesIndexing.time_getitem_pos_slice('string', 'nonunique_monotonic_inc')                                                                           |
| -        | 254±100μs                  | 148±2μs                            |    0.58 | indexing_engines.NumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.UInt64Engine'>, <class 'numpy.uint64'>), 'monotonic_decr', False, 100000)   |
| -        | 158±80μs                   | 90.6±0.9μs                         |    0.57 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc((<class 'pandas._libs.index.MaskedFloat32Engine'>, 'Float32'), 'non_monotonic', True, 2000000)                |
| -        | 169±80μs                   | 92.1±2μs                           |    0.54 | indexing_engines.MaskedNumericEngineIndexing.time_get_loc_near_middle((<class 'pandas._libs.index.MaskedInt32Engine'>, 'Int32'), 'non_monotonic', True, 2000000)        |
```

</DETAILS>

The result seem to vary everytime even though I've tried in the same commit. I also couldn't find any codes that call `_getitem_lowerdim()` inside `Index.get_loc()`. Did I get something wrong? Can the result vary depending on the PC's internal status or the surrounding environment?